+++ CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [0.3.0] - 2023-03-12
+### Changed
+- [with_annotated_args][tanchan.doc_parse.with_annotated_args] will now also parse
+  the docstring of the typed dict being used as an unpacked `**kwargs` type hint
+  for option descriptions.
+
 ## [0.2.2.post] - 2023-03-04
 ### Fixed
 - [SlashCommandGroup.as_sub_command][tanchan.doc_parse.SlashCommandGroup.as_sub_command]